test(coin_select): add inner prop test methods
