--- nursery/coin_select/tests/metrics.proptest-regressions
@@ -1,12 +1,0 @@
-# Seeds for failure cases proptest has generated in the past. It is
-# automatically read and these particular cases re-run before any
-# novel cases are generated.
-#
-# It is recommended to check this file in to source control so that
-# everyone who runs the test benefits from these saved cases.
-cc 78e8456749053271949d1821613de18d007ef6ddabc85eb0b9dc64b640f85736 # shrinks to n_candidates = 11, target_value = 500, base_weight = 0, min_fee = 0, feerate = 72.77445, feerate_lt_diff = 0.0, drain_weight = 100, drain_spend_weight = 1, drain_dust = 100
-cc 36b9844f4bd28caa412b4a7e384c370bf9406dd6d1cd3a37409181c096a3da95 # shrinks to n_candidates = 8, target_value = 378748, base_weight = 245, min_fee = 0, feerate = 90.57628, feerate_lt_diff = 41.46504, drain_weight = 408, drain_spend_weight = 1095, drain_dust = 100
-cc 9c5c20afb83a7b1b8dc66404c63379f12ac796f5f23e04ccb568778c84230e18 # shrinks to n_candidates = 11, target_value = 434651, base_weight = 361, min_fee = 0, feerate = 41.85748, feerate_lt_diff = 0.0, drain_weight = 100, drain_spend_weight = 1, drain_dust = 100
-cc 858be736b81a2b1ca5dafc2d6442c7facfd46af6d14659df3772daf0940b105e # shrinks to n_candidates = 3, target_value = 422791, base_weight = 272, min_fee = 0, feerate = 93.71708, feerate_lt_diff = 8.574516, drain_weight = 100, drain_spend_weight = 703, drain_dust = 100
-cc d643d1aaf1d708ca2a7ce3bf5357a14e82c9d60935b126c9b3f338a9bb0ebed3 # shrinks to n_candidates = 10, target_value = 381886, base_weight = 684, min_fee = 0, feerate = 72.56796, feerate_lt_diff = 0.0, drain_weight = 100, drain_spend_weight = 354, drain_dust = 100
-cc 931d5609471a5575882a8b2cb2c45884330cb18e95368c89a63cfe507a7c1a62 # shrinks to n_candidates = 10, target_value = 76204, base_weight = 71, min_fee = 0, feerate = 72.3613, feerate_lt_diff = 0.0, drain_weight = 100, drain_spend_weight = 357, drain_dust = 100
